Posibilidad de eliminar un postulante becario al rechazarlo.

+++ app/Http/Controllers/CompartidoDirecCoordController.php
@@ -14,10 +14,10 @@ use avaa\Alerta;
 use avaa\Solicitud;
 use avaa\Documento;
 use avaa\Imagen;
+use avaa\BecarioEntrevistador;
 use avaa\Events\SolicitudesAlerts;
 use Illuminate\Support\Facades\Auth;
 use DateTime;
-use avaa\BecarioEntrevistador;
 use avaa\Concurso;
 use PHPMailer\PHPMailer\PHPMailer;
 use PHPMailer\PHPMailer\Exception;
@@ -168,48 +168,51 @@ class CompartidoDirecCoordController extends Controller
             $id = $postulanteBecario->user->id;
             //borrar documentos
             
-            $fotografia = Imagen::where('user_id','=',$id)->where('titulo','=','fotografia')->first();
-            $cedula = Imagen::where('user_id','=',$id)->where('titulo','=','cedula')->first();
-            $constancia_cnu = Documento::where('user_id','=',$id)->where('titulo','=','constancia_cnu')->first();
-            $calificaciones_bachillerato = Documento::where('user_id','=',$id)->where('titulo','=','calificaciones_bachillerato')->first();
-            $constancia_aceptacion = Documento::where('user_id','=',$id)->where('titulo','=','constancia_aceptacion')->first();
-            $constancia_estudios = Documento::where('user_id','=',$id)->where('titulo','=','constancia_estudios')->first();
-            $calificaciones_universidad = Documento::where('user_id','=',$id)->where('titulo','=','calificaciones_universidad')->first();
-            $constancia_trabajo = Documento::where('user_id','=',$id)->where('titulo','=','constancia_trabajo')->first();
-            $declaracion_impuestos = Documento::where('user_id','=',$id)->where('titulo','=','declaracion_impuestos')->first();
-            $recibo_pago = Documento::where('user_id','=',$id)->where('titulo','=','recibo_pago')->first();
-            $referencia_profesor1 = Documento::where('user_id','=',$id)->where('titulo','=','referencia_profesor1')->first();
-            $referencia_profesor2 = Documento::where('user_id','=',$id)->where('titulo','=','referencia_profesor2')->first();
-            $ensayo = Documento::where('user_id','=',$id)->where('titulo','=','ensayo')->first();
-            
-            File::delete(substr($fotografia->url,1));
-            File::delete(substr($cedula->url,1));
-            File::delete(substr($constancia_cnu->url,1));
-            File::delete(substr($calificaciones_bachillerato->url,1));
-            File::delete(substr($constancia_aceptacion->url,1));
-            File::delete(substr($constancia_estudios->url,1));
-            File::delete(substr($calificaciones_universidad->url,1));
-            File::delete(substr($constancia_trabajo->url,1));
-            File::delete(substr($declaracion_impuestos->url,1));
-            File::delete(substr($recibo_pago->url,1));
-            File::delete(substr($referencia_profesor1->url,1));
-            File::delete(substr($referencia_profesor2->url,1));
-            File::delete(substr($ensayo->url,1));
 
-            $fotografia->delete();
-            $cedula->delete();
-            $constancia_cnu->delete();
-            $calificaciones_bachillerato->delete();
-            $constancia_aceptacion->delete();
-            $constancia_estudios->delete();
-            $calificaciones_universidad->delete();
-            $constancia_trabajo->delete();
-            $declaracion_impuestos->delete();
-            $recibo_pago->delete();
-            $referencia_profesor1->delete();
-            $referencia_profesor2->delete();
-            $ensayo->delete();
-            
+            $fotografia = Imagen::where('user_id','=',$id)->where('titulo','=','fotografia')->first();
+            if($fotografia)
+            {
+                $cedula = Imagen::where('user_id','=',$id)->where('titulo','=','cedula')->first();
+                $constancia_cnu = Documento::where('user_id','=',$id)->where('titulo','=','constancia_cnu')->first();
+                $calificaciones_bachillerato = Documento::where('user_id','=',$id)->where('titulo','=','calificaciones_bachillerato')->first();
+                $constancia_aceptacion = Documento::where('user_id','=',$id)->where('titulo','=','constancia_aceptacion')->first();
+                $constancia_estudios = Documento::where('user_id','=',$id)->where('titulo','=','constancia_estudios')->first();
+                $calificaciones_universidad = Documento::where('user_id','=',$id)->where('titulo','=','calificaciones_universidad')->first();
+                $constancia_trabajo = Documento::where('user_id','=',$id)->where('titulo','=','constancia_trabajo')->first();
+                $declaracion_impuestos = Documento::where('user_id','=',$id)->where('titulo','=','declaracion_impuestos')->first();
+                $recibo_pago = Documento::where('user_id','=',$id)->where('titulo','=','recibo_pago')->first();
+                $referencia_profesor1 = Documento::where('user_id','=',$id)->where('titulo','=','referencia_profesor1')->first();
+                $referencia_profesor2 = Documento::where('user_id','=',$id)->where('titulo','=','referencia_profesor2')->first();
+                $ensayo = Documento::where('user_id','=',$id)->where('titulo','=','ensayo')->first();
+                
+                File::delete(substr($fotografia->url,1));
+                File::delete(substr($cedula->url,1));
+                File::delete(substr($constancia_cnu->url,1));
+                File::delete(substr($calificaciones_bachillerato->url,1));
+                File::delete(substr($constancia_aceptacion->url,1));
+                File::delete(substr($constancia_estudios->url,1));
+                File::delete(substr($calificaciones_universidad->url,1));
+                File::delete(substr($constancia_trabajo->url,1));
+                File::delete(substr($declaracion_impuestos->url,1));
+                File::delete(substr($recibo_pago->url,1));
+                File::delete(substr($referencia_profesor1->url,1));
+                File::delete(substr($referencia_profesor2->url,1));
+                File::delete(substr($ensayo->url,1));
+
+                $fotografia->delete();
+                $cedula->delete();
+                $constancia_cnu->delete();
+                $calificaciones_bachillerato->delete();
+                $constancia_aceptacion->delete();
+                $constancia_estudios->delete();
+                $calificaciones_universidad->delete();
+                $constancia_trabajo->delete();
+                $declaracion_impuestos->delete();
+                $recibo_pago->delete();
+                $referencia_profesor1->delete();
+                $referencia_profesor2->delete();
+                $ensayo->delete();
+            }
             return response()->json(['success'=>'El Postulante ha sido Rechazado Exitosamente']);
        }
 
@@ -666,10 +669,37 @@ class CompartidoDirecCoordController extends Controller
     {
         $usuario = User::find($id); 
         $becario = Becario::find($id);
-
+        //Elimino la foto de perfil
         $img_perfil = Imagen::where('user_id','=',$id)->where('titulo','=','img_perfil')->first();
+        if($img_perfil)
+        {
+            if($img_perfil->url!=null)
+            {
+                File::delete(substr($img_perfil->url,1));
+            }
+            $img_perfil->delete();
+        }
+        
 
-        File::delete(substr($img_perfil->url,1));
-        return response()->json(['success'=>'El Postulante ha sido Aprobado Exitosamente']);
+        //Eliminio el documeto final entrevista
+        if($becario->documento_final_entrevista!=null)
+        {
+            File::delete($becario->documento_final_entrevista);
+        }
+
+        //Eliminado el documento que subió cada entrevistador
+        $entrevistadores = BecarioEntrevistador::paraBecario($id)->get();
+        foreach ($entrevistadores as $entrevistador)
+        {
+            if($entrevistador->documento!=null)
+            {
+                File::delete(substr($entrevistador->documento,1));
+            }
+        }
+        //Elimino el usuario  y el becario
+        $becario->delete();
+        $usuario->delete();
+        
+        return response()->json(['success'=>'El postulante ha sido eliminado exitosamente.']);
     }
 }
